model: add a basic model class
This class can help the user create his own model class.
It automatically reads lookup tables and lookup-table equations
if they are available.

It also provides some convenient functions such as
set_input_variables.
This enables the user to manually calculate certain input variables.
It is convenient if this happens after reading input files.
This requires some small adjustment of the original SimulationProblem
update function, but this requires a lot of copy-pasting.
A better solution should be found in the future.