model: add a basic class model class
This class can help the user create his own model class.

It can also be used as input argument to functions that extend
or complete models and thereby help reduce the number of lines
the user has to write.
For example, by passing a model object to a function,
a user does not have to collect relevant variables and files explicitly.